+++ src/utils/train_utils.py
@@ -45,7 +45,7 @@ def train_step(model: torch.nn.Module,
             # Frist convert to numpy array
             data, target = np.array(data), np.array(target)
             # Then convert to torch tensor
-            data, target = torch.from_numpy(data), torch.from_numpy(target)
+            data, target = torch.from_numpy(data).float(), torch.from_numpy(target).float() # .float() to ensure data type is float32 not 64
         data, target = data.to(device), target.to(device)
 
         # 1. Forward pass
